collection view: rename long file paths to avoid errors
