+++ src/packages/core/property-editor/uis/collection-view/config/bulk-action-permissions/collection-view-permissions.element.ts
