+++ src/packages/core/property-editor/uis/collection-view/config/bulk-action-permissions/manifests.ts
@@ -4,7 +4,7 @@ export const manifest: ManifestPropertyEditorUi = {
 	type: 'propertyEditorUi',
 	alias: 'Umb.PropertyEditorUi.CollectionView.BulkActionPermissions',
 	name: 'Collection View Bulk Action Permissions Property Editor UI',
-	element: () => import('./property-editor-ui-collection-view-permissions.element.js'),
+	element: () => import('./collection-view-permissions.element.js'),
 	meta: {
 		label: 'Collection View Bulk Action Permissions',
 		icon: 'icon-autofill',
