+++ src/packages/core/property-editor/uis/collection-view/config/column-configuration/collection-view-column-configuration.element.ts
