+++ src/packages/core/property-editor/uis/collection-view/config/column-configuration/manifests.ts
@@ -4,7 +4,7 @@ export const manifest: ManifestPropertyEditorUi = {
 	type: 'propertyEditorUi',
 	alias: 'Umb.PropertyEditorUi.CollectionView.ColumnConfiguration',
 	name: 'Collection View Column Configuration Property Editor UI',
-	element: () => import('./property-editor-ui-collection-view-column-configuration.element.js'),
+	element: () => import('./collection-view-column-configuration.element.js'),
 	meta: {
 		label: 'Collection View Column Configuration',
 		icon: 'icon-autofill',
