+++ src/packages/core/property-editor/uis/collection-view/config/layout-configuration/collection-view-layout-configuration.element.ts
