+++ src/packages/core/property-editor/uis/collection-view/config/layout-configuration/collection-view-layout-configuration.stories.ts
@@ -1,8 +1,8 @@
 import type { Meta, Story } from '@storybook/web-components';
-import type { UmbPropertyEditorUICollectionViewLayoutConfigurationElement } from './property-editor-ui-collection-view-layout-configuration.element.js';
+import type { UmbPropertyEditorUICollectionViewLayoutConfigurationElement } from './collection-view-layout-configuration.element.js';
 import { html } from '@umbraco-cms/backoffice/external/lit';
 
-import './property-editor-ui-collection-view-layout-configuration.element.js';
+import './collection-view-layout-configuration.element.js';
 
 export default {
 	title: 'Property Editor UIs/Collection View Layout Configuration',
