+++ src/packages/core/property-editor/uis/collection-view/config/layout-configuration/collection-view-layout-configuration.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { UmbPropertyEditorUICollectionViewLayoutConfigurationElement } from './property-editor-ui-collection-view-layout-configuration.element.js';
+import { UmbPropertyEditorUICollectionViewLayoutConfigurationElement } from './collection-view-layout-configuration.element.js';
 import { type UmbTestRunnerWindow, defaultA11yConfig } from '@umbraco-cms/internal/test-utils';
 
 describe('UmbPropertyEditorUICollectionViewLayoutConfigurationElement', () => {
