+++ src/packages/core/property-editor/uis/collection-view/config/layout-configuration/manifests.ts
@@ -4,7 +4,7 @@ export const manifest: ManifestPropertyEditorUi = {
 	type: 'propertyEditorUi',
 	alias: 'Umb.PropertyEditorUi.CollectionView.LayoutConfiguration',
 	name: 'Collection View Column Configuration Property Editor UI',
-	element: () => import('./property-editor-ui-collection-view-layout-configuration.element.js'),
+	element: () => import('./collection-view-layout-configuration.element.js'),
 	meta: {
 		label: 'Collection View Layout Configuration',
 		icon: 'icon-autofill',
